split jsonLoad.js and courseDB.js out of server.js
